innovation-catalyst-ai-agent/requirements.txt
# Core Dependencies
smolagents>=0.3.0
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pandas>=2.0.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
python-magic>=0.4.27

# NLP Libraries
spacy>=3.6.0
nltk>=3.8.0
scikit-learn>=1.3.0

# Web Interface
gradio>=3.40.0
streamlit>=1.25.0

# Utilities
pydantic>=2.0.0
loguru>=0.7.0
tqdm>=4.65.0
click>=8.1.0

# Development Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Optional Dependencies
faiss-cpu>=1.7.4  # For vector similarity search
plotly>=5.15.0    # For visualization


# Add to existing requirements.txt
python-dotenv>=1.0.0
huggingface-hub>=0.20.0

azure-ai-inference

smolagents>=0.3.0
litellm>=1.0.0
gradio>=4.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
PyPDF2>=3.0.0
python-docx>=0.8.11
spacy>=3.6.0
nltk>=3.8.0
scikit-learn>=1.3.0
scipy>=1.10.0
pydantic>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
